--- torchflare/callbacks/notifiers/__init__.py
@@ -1,4 +1,0 @@
-"""Imports for notifiers."""
-from torchflare.callbacks.notifiers.message_notifiers import DiscordNotifierCallback, SlackNotifierCallback
-
-__all__ = ["DiscordNotifierCallback", "SlackNotifierCallback"]
